feat: convert to python 3
